iniciando projeto de distancia minima baseado em divisao e conquista sh

+++ copiaMatriz.c
@@ -1,8 +1,8 @@
 #include "libsTrabalho.h"
-float **copiaMatrizImagem(float **mat, int lin, int col)
+double **copiaMatrizImagem(double **mat, int lin, int col)
 {
   int i, j;
-  float **matcopia = alocaMatrizImagem(lin, col);
+  double **matcopia = alocaMatrizImagem(lin, col);
   for (i = 0; i < lin; i++)
   {
     for (j = 0; j < col; j++)
