+++ imprimir.c
@@ -1,10 +1,10 @@
-void Imprimir(int *A, int tamanho)
+void Imprimir(double *A, int tamanho)
 {
     int i;
     printf("Vetor:\n");
     for (i = 0; i < tamanho; i++)
     {
-        printf("%d ", A[i]);
+        printf("%lf ", A[i]);
     }
     printf("\n");
 }
